Move Gauss-trick layers into `complextorch.nn.gauss` subpackage
Renames the hand-rolled real/imag-split conv/linear implementations out of
the top-level `complextorch.nn` surface and into a `complextorch.nn.gauss`
subpackage. The original `Slow*` prefix was a misleading legacy from when
these were *faster* than the naive split — since PyTorch 2.1.0's native
complex kernels they're slower than the `dtype=torch.cfloat` wrappers, and
the prefix described an accident of history rather than what the layers
actually do.

Renames (breaking, `2.0.0`):

- `complextorch.nn.SlowLinear`                → `complextorch.nn.gauss.Linear`
- `complextorch.nn.SlowConv{1,2,3}d`          → `complextorch.nn.gauss.Conv{1,2,3}d`
- `complextorch.nn.SlowConvTranspose{1,2,3}d` → `complextorch.nn.gauss.ConvTranspose{1,2,3}d`

The native-cfloat `Linear` / `Conv{1,2,3}d` / `ConvTranspose{1,2,3}d` keep
their top-level names. `complextorch.nn.gauss` is registered as a
subpackage alongside `init`, `relevance`, `masked`, `utils` — the prior
commit already wired the import into `complextorch/nn/__init__.py`; this
commit adds the subpackage itself plus the rest of the rename.

Internal callers:

- `MaskedChannelAttention{1,2,3}d` (six `conv_down` / `conv_up` sites in
  `attention/mca.py`) switched from `cvnn.SlowConv*d` to the native
  `cvnn.Conv*d`. These are 1×1 convs that gain nothing from the Gauss
  trick — the original choice was a holdover.

Tests:

- `tests/nn/gauss/{test_linear,test_conv}.py` — new files mirroring the
  package layout convention, with the forward / weight-property /
  `bias=False` cases moved from `tests/nn/modules/`.
- `tests/nn/modules/test_{conv,linear}.py` — stripped of `Slow*` blocks;
  now cover only the native cfloat wrappers.
- `tests/invariants/test_fast_slow_equivalence.py` →
  `test_native_gauss_equivalence.py` with imports and helper names
  updated. The property tests continue to assert that the two paths agree
  to floating-point tolerance on shared weights.

Docs:

- `CLAUDE.md` "Fast vs. Slow modules" section retitled to "Native vs.
  Gauss-trick modules" and rewritten to point at the new subpackage; the
  legacy `Slow*` naming is called out so future readers understand the
  history.

Git-rename caveat: because `modules/conv.py` and `modules/linear.py` were
*split* (Gauss-trick content into `gauss/`, fast wrappers stayed put),
git's default rename detection (≥50% similarity) sees them as "modified"
rather than "renamed". Use `git log --follow -C30%` or `git blame -C` to
trace the Gauss-trick content's history into its new location.
