+++ tests/nn/modules/test_conv.py
@@ -1,8 +1,7 @@
-"""Tests for complex Conv* and SlowConv* layers."""
+"""Tests for complex Conv* and ConvTranspose* layers (native cfloat wrappers)."""
 
 from __future__ import annotations
 
-import pytest
 import torch
 
 from complextorch.nn.modules.conv import (
@@ -12,12 +11,6 @@ from complextorch.nn.modules.conv import (
     ConvTranspose1d,
     ConvTranspose2d,
     ConvTranspose3d,
-    SlowConv1d,
-    SlowConv2d,
-    SlowConv3d,
-    SlowConvTranspose1d,
-    SlowConvTranspose2d,
-    SlowConvTranspose3d,
 )
 
 
@@ -64,112 +57,4 @@ def test_convtranspose3d_forward_shape():
     layer = ConvTranspose3d(2, 4, kernel_size=3, stride=2)
     x = torch.randn(1, 2, 2, 2, 2, dtype=torch.cfloat)
     out = layer(x)
-    assert out.is_complex()
-
-
-# ---------- Slow variants: forward + weight/bias properties ----------
-
-
-@pytest.mark.parametrize(
-    "cls, x_shape, kwargs",
-    [
-        (
-            SlowConv1d,
-            (2, 2, 8),
-            dict(in_channels=2, out_channels=4, kernel_size=3, padding=1),
-        ),
-        (
-            SlowConv2d,
-            (2, 2, 6, 6),
-            dict(in_channels=2, out_channels=3, kernel_size=3, padding=1),
-        ),
-        (
-            SlowConv3d,
-            (1, 2, 4, 4, 4),
-            dict(in_channels=2, out_channels=2, kernel_size=3, padding=1),
-        ),
-    ],
-)
-def test_slow_conv_forward(cls, x_shape, kwargs):
-    layer = cls(**kwargs)
-    x = torch.randn(*x_shape, dtype=torch.cfloat)
-    out = layer(x)
-    assert out.is_complex()
-    # weight/bias properties
-    assert layer.weight.is_complex()
-    assert layer.bias.is_complex()
-
-
-@pytest.mark.parametrize(
-    "cls, x_shape, kwargs",
-    [
-        (
-            SlowConv1d,
-            (2, 2, 8),
-            dict(in_channels=2, out_channels=4, kernel_size=3, padding=1, bias=False),
-        ),
-        (
-            SlowConv2d,
-            (1, 2, 6, 6),
-            dict(in_channels=2, out_channels=2, kernel_size=3, padding=1, bias=False),
-        ),
-    ],
-)
-def test_slow_conv_no_bias(cls, x_shape, kwargs):
-    layer = cls(**kwargs)
-    x = torch.randn(*x_shape, dtype=torch.cfloat)
-    out = layer(x)
-    assert layer.bias is None
-    assert out.is_complex()
-
-
-@pytest.mark.parametrize(
-    "cls, x_shape, kwargs",
-    [
-        (
-            SlowConvTranspose1d,
-            (2, 2, 4),
-            dict(in_channels=2, out_channels=3, kernel_size=3, stride=2),
-        ),
-        (
-            SlowConvTranspose2d,
-            (1, 2, 4, 4),
-            dict(in_channels=2, out_channels=2, kernel_size=3, stride=2),
-        ),
-        (
-            SlowConvTranspose3d,
-            (1, 2, 2, 2, 2),
-            dict(in_channels=2, out_channels=2, kernel_size=3, stride=2),
-        ),
-    ],
-)
-def test_slow_convtranspose_forward(cls, x_shape, kwargs):
-    layer = cls(**kwargs)
-    x = torch.randn(*x_shape, dtype=torch.cfloat)
-    out = layer(x)
-    assert out.is_complex()
-    assert layer.weight.is_complex()
-    assert layer.bias.is_complex()
-
-
-@pytest.mark.parametrize(
-    "cls, x_shape, kwargs",
-    [
-        (
-            SlowConvTranspose1d,
-            (2, 2, 4),
-            dict(in_channels=2, out_channels=3, kernel_size=3, bias=False),
-        ),
-        (
-            SlowConvTranspose2d,
-            (1, 2, 4, 4),
-            dict(in_channels=2, out_channels=2, kernel_size=3, bias=False),
-        ),
-    ],
-)
-def test_slow_convtranspose_no_bias(cls, x_shape, kwargs):
-    layer = cls(**kwargs)
-    x = torch.randn(*x_shape, dtype=torch.cfloat)
-    out = layer(x)
-    assert layer.bias is None
     assert out.is_complex()
